More consistent semantics for the vsg update algorithm.
At least, it's more consistent with my head cannon. If a change marker
is provided, use it to determine whether the GPU data should be updated
or not.  If it is not provided (AKA, NoChangeMarker) is used, then use
the overloaded == operator for the draw args to determine if the GPU
data needs to be updated.

If GPUData needs to be updated, update the GPU data.

Always merge the children with no shortcircuits.