feat!: More consistent semantics for the vsg update algorithm.  (#12)
* Directly check the type of drawargs and avoid two copies via tmp std::anys

* If you don't specify a change marker, expect it to re-create the part of the graph every time.

* More consistent semantics for the vsg update algorithm.

At least, it's more consistent with my head cannon. If a change marker
is provided, use it to determine whether the GPU data should be updated
or not.  If it is not provided (AKA, NoChangeMarker) is used, then use
the overloaded == operator for the draw args to determine if the GPU
data needs to be updated.

If GPUData needs to be updated, update the GPU data.

Always merge the children with no shortcircuits.

* Rename NoChangeMarker to NoChangeMarkerProvided which is clearer, IMO

* Fragments need to be comparable and properly dereference for this call

* Invert these conditions, which makes them correct

* Properly update the children

* Adding a Visibility class which hides things without removing them.

* Only use the equality operatore when they don't provide a change marker

* More sane checks

* Trigger CI

* :shipitparrot:

* Adding a TODO for review comments

---------

Co-authored-by: Lakin Wecker <lakin@wecker.ca>